demo/demo_classified/requirements.txt
django-classified==0.7.4
social-auth-app-django==2.1.0
